🚚 rename: event Form Type 이름 변경
