+++ src/components/openEvent/forms/index.ts
@@ -1,13 +1,13 @@
 // export { default as AdditionalInfo } from './AdditionalInfo/AdditionalInfo';
-// export { default as Address } from './Address/Address';
+export { default as Address } from './Address/Address';
 export { default as ApplicationPeriod } from './ApplicationPeriod/ApplicationPeriod';
-// export { default as Cost } from './Cost/Cost';
-// export { default as Description } from './Description/Description';
+export { default as Cost } from './Cost/Cost';
+export { default as Description } from './Description/Description';
 export { default as Field } from './Field/Field';
-// export { default as HostEmail } from './HostEmail/HostEmail';
-// export { default as HostName } from './HostName/HostName';
-// export { default as HostPhoneNumber } from './HostPhoneNumber/HostPhoneNumber';
+export { default as HostEmail } from './HostEmail/HostEmail';
+export { default as HostName } from './HostName/HostName';
+export { default as HostPhoneNumber } from './HostPhoneNumber/HostPhoneNumber';
 export { default as Period } from './Period/Period';
-// export { default as RecruitmentNumber } from './RecruitmentNumber/RecruitmentNumber';
+export { default as RecruitmentNumber } from './RecruitmentNumber/RecruitmentNumber';
 export { default as Title } from './Title/Title';
-// export { default as UploadImage } from './UploadImage/UploadImage';
+export { default as UploadImage } from './UploadImage/UploadImage';
